Steffen Wieschalla <steffen.wieschalla@student.uni-tuebingen.de> says:
I tried to complile ddd-2.1. configure files with the following error
message:

checking for XtToolkitInitialize in -lXt... no
configure: error: The X toolkit library '-lXt' could not be found.
                  Please use the configure options '--with-x-includes=DIR'
                  and '--with-x-libraries=DIR' to specify the X location.
                  See the file 'config.log' for further diagnostics.
configure: error: ./configure failed for ddd

The error is that you need to use --x-includes and --x-libraries
instead of --with-x-includes and --with-x-libraries


git-svn-id: svn://svn.savannah.gnu.org/ddd/trunk@1783 1d73a310-8f1f-0410-b01a-dab0232e19a9