Avi Cohen Stuart <avi@baan.nl> says:
Maybe we shouldn't run under root...
The problem that occurred was that the /bin and /lib were removed
on a HP-UX 10.20.
It appeared that these file (/bin is a symlink to /usr/bin and
/lib a symlink /usr/lib) disappeared after calling ddd.

The script responsable for this is found in xlibdir.C that is included
in xconfig.C.
The following line did it:

rm -f /INFORMIXTMP /Mail /SD_CDROM /actions /auzot1 /auzot2 /beta /bin
/cdrom /c mds /dbms /dev /etc /export /home /lib /lost+found /monitor
/net /nhome /objvobs /opt /sbin /srcvobs /stand /tmp /tmp_mnt /trash
/usr /usr1 /usr10 /usr11 /usr12 /usr13 /usr14 /usr2 /usr3 /usr4 /usr5
/usr6 /usr7 /usr8 /usr9 /validate /var /v iew /vobs /vobstore /.baan
/.brand /.elm /.lrom /.sw

It is generated from the script in xlibdir.C but we were not able to
reproduce this why it happened.

A workaround is the following diff of xlibdir.txt, I added an

export tmp


git-svn-id: svn://svn.savannah.gnu.org/ddd/trunk@2589 1d73a310-8f1f-0410-b01a-dab0232e19a9